Add testing framework
Add a testing framework, written in Rust, that tests the results of
syntax highlighting.

The test cases are in `tests/cases`. Each case is a file called
`NAME.just`, with a corresponding `NAME.html`. The test framework
applies syntax highlighting to `NAME.just`, and converts it to HTML
using `:TOhtml`. The output is diffed against `NAME.html`, and if the
diff is non-empty, the test fails.

A GitHub Actions workflow will also run the tests on every pull request.

Two small test cases are included, `tests/cases/recipe.just` and
`test/cases/comment.just`.