justfile: Make preview tasks a cleanroom, stock editor plus this plugin
There are currently quite a few ways for user configs to leak into this
and change how the plugin responds. Notably if the user has vim-polyglot
or anything else that does Justfile stuff to start with, these preview
tasts will not actually show what *this* plugin is doing.

This should be much more robust, including leaving the editor to
actually *want* to load the plugin as a result of normal actions like
how it does file type detection, not forcing the plugin to be loaded
before testing begins.